requirements.txt
bs4==0.0.1
langchain==0.0.237
requests==2.28.2
xmltodict==0.13.0
openai==0.27.0
tiktoken==0.3.0
faiss-cpu==1.7.3
chromadb
